Update fixtures for valid blueprints as per blueprints v2 spec (#214)
In my wpcom work, I found the following valid blueprint fixtures were
out of date as per blueprints v2 spec. Our validator flagged them, and
it now passes with the changes in this PR.

Also, note that these fixtures files are currently not used in the repo.
Nor they were in the original commit. Seems like an intermediary
artifact. With this PR, we can atleast have them to be upto dated & can
possibly start using them in tests that we can add. In #213 I have the
entire incorrect valid fixture file in the test, but if they are indeed
to kept around, it can also use the fixture file.